+++ lib/karafka/pro/processing/strategies/dlq/ftr.rb
@@ -41,7 +41,7 @@ module Karafka
 
                   mark_as_consumed(messages.last)
 
-                  throttle_or_seek_if_needed
+                  handle_post_filtering
                 elsif coordinator.pause_tracker.attempt <= topic.dead_letter_queue.max_retries
                   retry_after_pause
                 else