+++ lib/karafka/pro/processing/strategies/lrj/ftr_mom.rb
@@ -41,8 +41,8 @@ module Karafka
                   return if coordinator.manual_pause?
 
                   # If still not revoked and was throttled, we need to apply filtering logic
-                  if coordinator.throttled? && !revoked?
-                    throttle_or_seek_if_needed(resume_on_seek: true)
+                  if coordinator.filtered? && !revoked?
+                    handle_post_filtering
                   elsif !revoked?
                     # If not revoked and not throttled, we move to where we were suppose to and
                     # resume