+++ spec/lib/karafka/pro/processing/filters_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  subject(:filters) { described_class.new(filters_array) }
+
+  let(:filters_array) { [] }
+
+  pending
+end